Rename ics02_client::client module to ics02_client::raw
