+++ modules/src/ics02_client/mod.rs
@@ -1,9 +1,9 @@
 //! ICS 02: IBC Client implementation
 
-pub mod client;
 pub mod client_type;
 pub mod error;
 pub mod events;
 pub mod header;
 pub mod msgs;
+pub mod raw;
 pub mod state;
